tilelink2: refactor RegField into interface and implementation
